// src/types/DataTypes.ts
export default interface DataTypes {
  address?: string;
  benefits?: Array<string>;
  description?: string;
  email?: string;
  employment_type?: Array<string>;
  id?: string;
  name?: string;
  phone?: string;
  pictures?: Array<string>;
  salary?: string;
  title?: string;
  updatedAt?: string;
}